backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.36.0
openai>=1.57.0
python-dotenv>=1.0.0
pydantic>=2.10.0
json-repair>=0.30.0
kuzu
